% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get detailed info on R package datasets}
\usage{
get_dataset_info(packagenames = NULL, allclasses = FALSE)
}
\arguments{
\item{packagenames}{a character vector providing the package(s) to look in for data sets, or \code{NULL}. If \code{NULL}, all loaded packages will be searched.}

\item{allclasses}{a logical. If \code{TRUE} a column with all classes (not only the first one listed) is included. Defaults to \code{FALSE}.}
}
\description{
finds all datasets in a specified package and returns a tibble with information on the dimensions of the datasets as well as data types. It is an alternative to \code{data()} which simply lists the datasets with short descriptions.
}
\section{Output columns}{

\itemize{
\item \code{package} name of package
\item \code{name} name of dataset
\item \code{dim} dimensions of dataset or \code{NA} for vectors, lists
\item \code{length} length (vectors only)
\item \code{first_class_listed} first class listed
\item \code{n_cols} number of numeric columns
\item \code{i_cols} number of integer columns
\item \code{f_cols} number of factor columns
\item \code{c_cols} number of character columns
\item \code{d_cols} number of date columns
\item \code{other_cols} number of other columns
\item \code{allclasses} full list of classes \link{optional}
}
}

\examples{

x <- get_dataset_info("ggplot2")
View(x)

x <- get_dataset_info()
View(x)

x <- get_dataset_info(c("fivethirtyeight", "pgmm"))
View(x)
}
